ImageProfileTypeChecking

+++ src/chat/chat.service.ts
@@ -295,8 +295,10 @@ export class ChatService extends GenericsService<GroupChat, GroupChat, GroupChat
             .getOne() as any;
         profile = typeof profile === 'string' ? profile : profile?.id;
         // check if the profile exists
+        if (!groupChat)
+            throw new NotFoundException('Chat not found');
         if (profile && !groupChat.groupChatToProfiles.find(gctp => gctp.profile.id === profile))
-            throw new ForbiddenException('You are not in this group chat');
+            throw new ForbiddenException('You are a member of this chat');
         groupChat?.lastMessage?.fromJson();
         return groupChat;
     }
@@ -358,12 +360,16 @@ export class ChatService extends GenericsService<GroupChat, GroupChat, GroupChat
         return groupChat;
     }
 
-    async deleteGroupChat(groupChat: GroupChat | string): Promise<void> {
-        groupChat = typeof groupChat === 'string' ? await this.groupChatRepository.findOneByOrFail({ id: groupChat }) : groupChat;   
-        if(groupChat.isPrivate)
-            throw new ForbiddenException('You cannot delete a private chat');     
+    async deleteGroupChat(groupChat: GroupChat | string, profile: Profile | string): Promise<void> {
+        groupChat = typeof groupChat === 'string' ? await this.groupChatRepository.findOneByOrFail({ id: groupChat }) : groupChat;
+        const isUserAdmin = await this.isUserAdminOfGroupChat(groupChat, profile);
+        if (!isUserAdmin)
+            throw new ForbiddenException('You are not an admin of this group chat');
+        if (groupChat.isPrivate)
+            throw new ForbiddenException('You cannot delete a private chat');
         await this.groupChatRepository.delete(groupChat.id);
     }
+
     private deleteFiles(files: Attachment[]) {
         files.forEach(async file => {
             try {
